requirements.txt
daphne==3.0.1
Django==3.1.2
django-ckeditor==6.0.0
django-js-asset==1.2.2
Pillow==8.1.0
sqlparse==0.4.1
uvicorn==0.13.4
whitenoise==5.2.0
gunicorn==20.1.0